Fix #492: force-refresh when position or dimension changes.
This should resolve #492 in its entirety. Basically, as the data was
already loaded, the `shouldUpdate` function was returning false, meaning
that the "new" data coming in never replaced the old.

This led to the `stillValid` check going "wait no, we're miles away from
where we should be" and returning false, resulting in the chest
immediately closing.